functional tests: Move test aci server to testutils
