+++ tests/rkt_auth_test.go
@@ -22,8 +22,8 @@ import (
 	"strings"
 	"testing"
 
-	taas "github.com/coreos/rkt/tests/test-auth-server/aci"
 	"github.com/coreos/rkt/tests/testutils"
+	taas "github.com/coreos/rkt/tests/testutils/aci-server"
 )
 
 func TestAuthSanity(t *testing.T) {
